% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RITM2.R
\name{point_to_point}
\alias{point_to_point}
\title{Point to Point ITM}
\usage{
point_to_point(struct_Input)
}
\arguments{
\item{struct_Input}{Named list object of input parameters:
\itemize{
  \item Frequency - Frequency to calculate loss at (Hz)
  \item Elevation - terrain elevation profile, (list of points) (m)
  \item Resolution - terrain input resolution (distance b/t points) (m)
  \item TX_Height - Transmit antenna height above ground (m)
  \item RX_Height - Recieve antenna height above ground (m)
  \item eps - Soil dielectric
  \item sgm - Surface conductivity
  \item surfref - Surface refractivity
  \item Climate -  Climate, 1-Equitorial, 2-Continental Subtropical, 3-Maritime Tropical, 4-Desert
  \item Polarization - 1 is vertical, 0 is horizontal.
  \item Confidence - confidence for statistical analysis (.01 to .99)
  \item Reliability - Reliability to calculate statistics for (.01 to .99)
}}
}
\value{
Path loss (dB), error ID, error message, and mode.
}
\description{
Returns path loss in Longley Rice point-to-point mode. Best for rural areas.
}
\examples{

#commented below is an example of how to get an elevation profile in R.
#library(elevatr)
#library(sp)
#set.seed(65.7)
#examp_df <- data.frame(x = runif(3, min = -73, max = -72.5), y = runif(3, min = 42,  max = 43))
#prj_dd <- "+init=EPSG:4326"
#cats <- data.frame(category = c("H", "M", "L"))
#examp_df2 <- data.frame(examp_df, cats)
#examp_sp <- SpatialPoints(examp_df, proj4string = CRS(prj_dd))
#examp_spdf <- SpatialPointsDataFrame(examp_sp, data = cats)
#df_elev_epqs <- get_elev_point(examp_df, prj = prj_dd, src = "epqs")
#Elevation<-df_elev_epqs$elevation

#These are the values returned above:
Elevation<-c(207.81, 198.95, 306.15)

#Build the input list
struct_Input<-list()
struct_Input$Frequency<-120*1000000 #Frequency to calculate loss at (Hz)
struct_Input$Elevation<-Elevation #terrain elevation profile, (list of points) (m)
struct_Input$Resolution<-40000 #terrain input resolution (distance b/t points) (m)
struct_Input$TX_Height<-3 #Transmit antenna height above ground (m)
struct_Input$RX_Height<-100 #Recieve antenna height above ground (m)
struct_Input$eps<-15 #Soil dielectric
struct_Input$sgm<-.005 #Surface conductivity
struct_Input$surfref<-301 #Surface refractivity
struct_Input$Climate<-5
struct_Input$Polarization<-1 #1 is vertical, 0 is horizontal
struct_Input$Confidence<-.95 #confidence for statistical analysis
struct_Input$Reliability<-.95 #Reliability to calculate statistics for (.01 to .99)

point_to_point(struct_Input)
}
